requirements.txt
flask
gunicorn
psycopg2-binary
newrelic
requests
opentelemetry-sdk
opentelemetry-api
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-psycopg2
opentelemetry-exporter-otlp
prometheus-flask-exporter